third_party/blink/web_tests/pointer-lock/pointer-lock-api-expected.txt

Basic API existence test for Pointer Lock.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.onpointerlockchange is defined.
PASS document.onpointerlockerror is defined.
PASS document.exitPointerLock is defined.
PASS document.pointerLockElement is null
PASS element.requestPointerLock is defined.
PASS successfullyParsed is true

TEST COMPLETE
